[#58] Logical plans, physical plans, more robust query support
This fix includes many changes, and removes a fair amount of technical
debt that had been acquired during the sprint to handle basic queries.

Formally organizing plans into logical and physical plans presents a
better way to handle optimizations and more advanced query generation;
for example, various types of JOINs and UNIONs.

The emitted plans are now stored in files, rather than in globals,
allowing for easier inspection of the code to ensure quality. This
also simplifies a great deal of the CURSOR behavior, and allows more
edge cases to be handled with easier code.

Plans can be chained together to handle nested queries; this presents an
opportunity for parallelism as a future enhancement.

optimize_logical_plan presents an opportunity to explore multiple plans,
once we have statistics stored in the database to calculate plan cost.

Changes to CMake were done, as part of pulling in the YDBCmake repository.