[#46] Added make_error_response and test cases, small changes
Testing for make_error_response revealed an issue with htonl and the
integer type that was being used; a host-ordered integer at a small
number is a negative number when converted to network-order, which
presented issues when doing a comparison. Changing the datatypes fixed
this.

An extra byte of length was removed from various places, as the
message format is not counted in the length of the message.