[#50,#46] Move query logic to handle_query, prevent double free's
Previously there was just a mess of code in octo.c which read in a
string and passed it too emit_select_statement/emit_delete_statement
or whatever was appropriate; this logic was moved to it's own routine
so it can be called by the handle_* routines for the PSQL wire
protocol.

Additionally, fixed some issues with cleanup_sql_statement (most
noteably, missing statement types). Also adjusted it so each statement
copies the pointer for its type to a temporary variable, then sets the
inner statement pointer to NULL. This means that if
cleanup_sql_statement gets called multiple times for the same
statement, no harm will come of it. This should greatly reduce the
problems I was fearing regarding shared pointers between statements
and logical plans and what not, and allow us to be more efficient with
the number of copies we do at runtime.