Add Northwind database from https://www.w3schools.com/sql/sql_distinct.asp (will be used in later commits)
This commit contains files that load the Northwind database in Octo as well as in Postgres.
This will enable queries against the Northwind database to be run in Octo and cross-checked
against Postgres (in automated Octo bats tests).

The list of tables and records for each table were obtained from the above mentioned url
by running SQL queries `select * from <tablename>` in the browser. There were some columns
that contained non-ascii data and those were trimmed out to keep it simple.