[#127] Ensure that unique_id's are filled out for SELECT statements
Previously, Octo would assign unique_id's to top-level SELECT statements
during the logical planning phase. This becomes problematic if you
attempt to use a calculated column in the same query where it is being
defined. To make this doable, we need to assign unique_id's earlier on
in the process, which involves changing the return types from some
parser branches from SqlSelectStatement to SqlTableAlias, and updating
various other bits of code associated.