[#155] Ensure we check $DATA before following a KEY_FIX
When a key is optimized so that it's value is a constant (it is FIXED to
a particular value), we need to verify that the value exists as a valid
option, and if not, skip over that iteration. This was troublesome when
we had a cross reference key, which could be fixed, and was later used
in an expression against the value it fixed too.

This commit solves the issue by checking that there is data in the cross
reference index for the fixed value of the cross reference.