Fix generated cross-reference M program to maintain ^%ydboctoxref correctly
An M IF statement of the form "IF $INCREMENT(a,-1),$INCREMENT(b-1)" will not execute
$INCREMENT(b-1) if the $INCREMENT(a,-1) operation returns a 0. But what is intended
in the physical plan is to do both the $INCREMENTs or do neither. Therefore the fix
is to split the $INCREMENTS into separate lines each with their own IF check i.e.
	IF $INCREMENT(a,-1)
	IF $INCREMENT(b,-1)

A similar fix is done to the line of the form "IF $INCREMENT(a),$INCREMENT(b)" even
though it is not needed for correctness. It is done for consistency and readability
of the generated M code.

While at this, fixed the indents of a few dotted do levels and removed some trailing
blank space.

Later it was pointed out that a slightly better way to do the same thing is
	IF $SELECT(INCREMENT(a,-1):$INCREMENT(b,-1),1:$INCREMENT(b,-1))

So that is what is finally done even though it reduces the code readability a bit.