[#107] Fix memory issue in find_table, cleanup memory after SELECT
Calls to libocto functions in find_table were using octo_cmalloc, which
was placing a SqlStatement in a memory chunk that was later cleaned.
This resulted in strange behavior since it would get overwritten by
another valid-looking SqlStatement.

The fix was to allocate a new, temporary memory chunk to store the
decompressed table in.

Also, we call octo_cfree on memory chunks which were not being cleaned
after a SELECT statement ran. The trick was finding the right location
to call it, since several rocto routines used the SqlStatement after the
data was sent to the client.