Removed erroneous warnings caused by PostgreSQL protocol connection resets from Rocto log messages
The postgres wire protocol includes a connection reset after the initial StartupMessage or SSLRequest, which led to read_bytes reading 0 bytes, returning 1, and causing read_message return NULL, and rocto.c to issue the warning on seeing the NULL.

However, the warning is misleading, since this is expected behavior. As a result, I adjusted the relevant error check to prevent the warning from being issued and clarify that there is no actual error. I then manually confirmed and added a test case.