Performance improvements for some types of SQL queries
See test case "multiple inner joins across multiple tables that should optimize to ONE FOR loop across ONE table"
in tests/test_inner_join.bats.in for details on the query which ran very slow before the fixes in this
commit. That query which involved multiple joins resulted in M code which has FOUR levels of nested FOR
loops (i.e. not optimized at all) whereas with the optimization changes done in this commit the generated
M code now does just ONE FOR loop for the exact same query.

The main fix is to lp_optimize_where_multi_equals_ands() which is now reworked to do a recursive descent
in case of a LP_BOOLEAN_AND.

In addition what used to be key_unique_id_ordering[] is now renamed as key_unique_id_array[] and initialized
outside of the recursion and passed in as a parameter to avoid being reconstructed at each recursion level.
Towards this, a helper function lp_optimize_where_multi_equal_ands_helper() is now created.

A few test cases are also added to test the changes (the main one being tests/test_inner_join.bats.in).

An existing test reference file tests/outref/T0009.ref is fixed to reflect an additional trigger that
gets added due to additional optimization of the same query with the new changes.

Various existing test scripts have been enhanced to display the actual output too (in addition to storing
the output in a file called output.txt) in case a test fails.