[#177] Added support for query cancellation to Rocto
Previously, Rocto could not handle client requests to cancel running queries. This feature is now supported in compliance with the [PostgreSQL wire protocol](https://www.postgresql.org/docs/11/protocol-flow.html#id-1.10.5.7.9).

A number of changes were required to support this feature, including:

+ Introduction of a new signal handler in rocto.c for receiving and forwarding SIGUSR1 from processes handling cancel requests to running M query generation routine
+ Introduction of a new M routine for cleaning up partially generated cross references and triggers should a cancel request be received
+ Revisions to the rocto connection loop in rocto.c to prepare for and receive cancel requests, including issuing backend secret key data to new client connections
+ Addition of several new error and informational messages
+ Unit and integration tests for new functionality

Note that no code was added for acting on SIGUSR1 in Rocto C code, as no code blocks performed slowly enough to be testable/cancelable. This is liable to change in the future, and adjustments to some loops and/or recursive functions may be made as needed.

Additionally, the relevant BATS test has been temporarily disabled in the pipeline due to inconsistent, timing related failures that do not occur on development or test machines.