octo -vv in a terminal should display input query before it gets run (not before input of next query)
Previously, octo -vv in a terminal used to display the input query AFTER it has been executed.
It also used to display all queries input till now (including the entire history of query input).
Both these issues are fixed for the terminal case.

For a file, octo -vv continues to display the entire file contents as part of executing each query
in case the file contains multiple queries. This is not as easy to fix so is not done as part of
this commit. Will be addressed at a later point if/when the need arises.