As noted in issue #201, Octo previously allowed users to specify the prefix of Octo-related global variables using a setting in a configuration file. This is unnecessary and has therefore been removed.
Specifically, the octo_global_prefix setting was removed from the configuration file and octo_init no longer checks for this setting. Additionally, all functions that relied on this setting now utilize a fixed string, "%ydbocto", to specify the global prefix.

No tests were added as this is a minor implementation change that is difficult to test. It is sufficient for existing tests to pass/not break.