[#225] Add session cleanup to rocto.
As noted in issue #225, rocto previously did not cleanup session information when clients disconnected. Now each rocto session is cleaned up when clients disconnect.

This was accomplished by modifying rocto.c to delete session information from the global session variable based on the session id. Error checks were also added along with an INFO message when the cleanup is successful.

A test was added to verify that successful deletion occurs on session close.