[#178] Added "-h" and "--help" command line options to octo and rocto executables.
This was accomplished by adding a usage strings for octo and rocto, respectively, and an additional option set for handling rocto option in getopt_long.

Handling for unknown and invalid options was modified to suggest using the help option to the user and then exit with status of 1. Additionally, the -f and -d options are now invalid for rocto and are handled accordingly.

Tests added for the above changes.