[#248] Log generated M file name (for SQL query) in octo -vv and rocto -vv output
Each SQL query generates one M file name of the form _ydboctoP*.m (P stands for Plan)
and one or more M files of the form _ydboctoX*.m (X stands for Cross Reference) depending
on how many cross references need to be built.

octo -vv and rocto -vv now log the generated M file name in both cases. In the case
of _ydboctoP*.m, the M file name is printed even in case it is already present. Whereas
in the _ydboctoX*.m case, the M file name is printed only in case it is generated as part
of the current query. While it would be nice to print the M file name in the latter case too
it is not clear how straightforward it is since emit_physical_plan() (which is where the
_ydboctoX*.m file name is accessible) is not even called in case _ydboctoP*.m file exists.
Will address this edge case later if it is seen as really desirable.