Fix SIG-11 in rocto when stacksize limit is low and rocto is built with huge STRING_BUFFER_LENGTH
If the C-stacksize limit is set to a low value (say 8Mb) and rocto is built by passing a huge buffer
size (e.g. `-DSTRING_BUFFER_LENGTH=600000`) to the cmake command (normally done for the Vista Octo
database environment), executing any query in rocto will cause it to terminate with a SIG-11. The
workaround till now was to increase the stacksize of the process to a high value (close to 800Mb or so)
in the shell (using the limit or ulimit command) before starting rocto. This is no longer necessary.

We were allocating a huge array in the C-stack in handle_query_response() as part of the following line.

```
DataRowParm data_row_parms[MAX_STR_CONST];
```

`MAX_STR_CONST` is the same value as `STRING_BUFFER_LENGTH` specified in the cmake command. So this meant
allocating 600,000 structures of type `DataRowParm` each of size 16 bytes. Resulting in a C-stack size
of more than 8Mb which causes the first invocation of handle_query_response() (i.e. running any query
through rocto) to result in a SIG-11. This is now reworked to allocate only 16 structures at the start
and dynamically reallocating this array as the need arises.