Make recently added test_psql_authentication.bats test pass in a non-pipeline environment
In a non-pipeline environment, I encountered failures in 2 subtests in the psql_authentication test.
This was recently introduced. My test run failed in the following secondary error check.

   `[[ "$output" =~ "psql: could not connect to server: Connection refused" ]]`

No idea why it passes in the pipeline though. The real point of the failing 2 subtests is to check
the following (primary error check) which continues to pass.

   `[[ "$output" =~ "password doesn't match stored value" ]]`
   `[[ "$output" =~ "failed to retrieve user info from database" ]]`

Maybe the version of psql is different and so the output is slightly different.
In any case, the primary error check is still there. The secondary error check is now removed
and instead we also check that the exit status is non-zero (equals 2).