[#215] SELECT xxx FROM TABLE dumps core if xxx is not a valid column in TABLE
Problem statement
------------------
SELECT xxx FROM TABLE used to issue an `Unknown column: xxx` error previously if xxx is not a valid
column in TABLE but as part of a prior commit (SHA d373060e12ea99d84ff2e3daf319ca90cabce449), this
stopped working and instead dumped core due to a recursion in populate_data_type() resulting in a
C-stack overflow.

That prior commit enabled the following query to compile fine where the `order by` clause was
referring to a column `abcd` which was not an actual column but an alias column obtained by using
the `as` clause (corresponding to the actual column `firstname`).

`select firstname as abcd from names order by abcd`

But that also caused queries like the following to behave differently.

`select invalid as names`

This is because the column name `invalid` assumed an alias of `invalid` (the same name) even
though none was specified by the user in an `as` clause. And so later when we tried to see if
`invalid` is a valid column name we find it while going through the list of aliased column names
at the end of `qualify_column_name()` and incorrectly decide it is a valid column name. This causes
`qualify_query()` (which is what eventually calls `qualify_column_name()`) to return success which
then causes us to continue with processing the query (and eventually go to `populate_data_type()`
where the C-stack explosion occurs) instead of exiting the parse right there.

Fix
----
Once an entire query is parsed, we invoke `qualify_query()` to check various aspects of the query,
including if all column names specified are valid.

In `qualify_query()`, we check the select_list (list of columns in the SELECT clause) first
by invoking `qualify_statement(select->select_list, ...)`. We later check the where clause in
similar fashion `qualify_statement(select->where_expression, ...)`. And finally we check the order
clause `qualify_statement(select->order_expression, ...)`. Each `qualify_statement()` call does
validation of any columns specified in those clauses. We do strict verification of columns for all
`qualify_statement()` invocations except the last one (`select->order_expression` invocation). In
that case, we do relaxed column validation.

This is done by passing in the parameter MATCH_QUALIFIED_COLUMNS_TRUE (i.e. TRUE) for the strict
validation case and MATCH_QUALIFIED_COLUMNS_FALSE (i.e. FALSE) for the relaxed validation case. This
also meant passing this parameter down various `qualify_*()` function calls that are invoked
recursively by `qualify_statement()`. The full list of those files whose interface changed is

	match_column_in_table.c
	qualify_column_list.c
	qualify_column_list_alias.c
	qualify_column_name.c
	qualify_statement.c

This parameter is finally used in `qualify_column_name.c` and `match_column_in_table.c`. If this
parameter is TRUE, a strict column validation is done. This means the input column name is considered
qualified only if it matches a previously qualified column name. And a previously qualified column
name is identified by the fact that in the `SqlColumnListAlias` structure, the `column_list` member
is of type `column_alias_STATEMENT` (`qualify_statement()` resets the type from `value_STATEMENT`
to `column_alias_STATEMENT` in the valid column name case).

`match_column_in_table()` is also called by `natural_join_condition()`. That call was changed to
pass MATCH_QUALIFIED_COLUMNS_FALSE as the new parameter since we do not want strict column name
validation in that case (we have not yet reached `qualify_query()` stage).

Miscellaneous changes
----------------------
* A memcmp in `qualify_column_name.c` did not also check the length of the two strings for equality
  which meant if one string is a substring of the other, the memcmp will return 0 incorrectly. That
  is fixed to additionally do the length equality check.

* Various octo_cmalloc() calls that allocated one structure have been recoded using a new
  OCTO_CMALLOC_STRUCT macro that makes the code a lot more readable as well as gives one
  centralized control in the macro (e.g. add fprintf code to see all structure allocations
  that happen for a given query etc.). Because of this, another related macro MALLOC_LP was
  changed to have a new parameter (first parameter, which was previously used as the result
  of the MALLOC_LP macro invocation) and that also meant adding another MALLOC_LP_2ARGS macro
  which was needed in case the return value was not assigned to any variable. These changes
  touched a lot of the modules.

* A boolean_t type is now introduced.

* A file `qualify_join_conditions.c` was unused so it is removed.

* Various test cases have been introduced to test invalid column names specified in queries,
  sub-queries etc. The test_select_columns.bats.in file is heavily reworked to add these tests.