[#260] Added COLUMNS environment variable to prevent erroneous Octo test output due to long lines.
When Octo output from both stdout and stderr is redirected using "&>" and the query used is too long, the output gets mangled leading to inaccurate outrefs or broken tests.

This is due to an insufficient number of terminal columns being available to contain the query. The fix is to set the COLUMNS environment variable sufficiently large to represent the queries under test on a single line.