[#161] Fix octo ignoring the next query following a query that used an unknown column in a clause (as, where, etc)
Previously the parser would call YYERROR if an unknown column was used
in a clause. YYERROR however attempts to recover the parse and read more.
This means that the next query would not be run. Now the parser calls
YYABORT which stops the parse and returns to run_query. This is the
approprate action for octo to take when encountering a parser error.

This commit also modifies yyerror.c to properly format multiple queries
on the same line. And makes octo emit a newline when exiting with EOF.
A large number of tests needed to have a newline appened to the end of
the reference file to account for this.