[#267] Ctrl-C of rocto now terminates the process
rocto was using signal() to establish the signal handler for SIGINT (Ctrl-C).  The man pages of
signal() clearly indicate that to establish a signal handler one needs to use sigaction() and not
signal(). Most likely this is why Ctrl-C worked on some environments and did not on some other
environments.

After changing signal() to instead use sigaction(), Ctrl-C works on at least one environment where
it was reliably not working previously.