[#224] SIG-11 when one of JOIN tables is created by a sub-query that references a table from the outer query in the JOIN
See https://gitlab.com/YottaDB/DBMS/YDBOcto/issues/224 for the issue description.

The SIG-11 happened because we let the query parse fine and went ahead to generate code.  But a JOIN
table that involves a sub-query should not reference any table names from the parent query. This
should have been signaled as an error (invalid tablename/column etc.).  This is now done in
`qualify_query()` where we pass NULL as the `parent_join` parameter when we try to validate each
of the join tables in the current query. This means the parent tables will not be added (using
`dqinsert`) to the list of join tables in the nested invocation of `qualify_query()`.

Note that an "Unknown column" error is now issued in this case instead of the previous SIG-11.
The error message could be enhanced to indicate this is an invalid reference to the parent table name
in a sub-query. But that requires more effort and we will worry about that at a later point in time.

A new bats test has been added for this (filename test_subquery_invalid_table.bats.in).