[#236] Change ^%ydboctocursor and ^%ydboctosession usages from global variables to local variables (i.e. to %ydboctocursor and %ydboctosession)
This is expected to slightly improve performance. This change was possible since neither of those globals
currently store persistent data (i.e. data that is required by anyone other than the current process).

Various src/m_templates/*.ctemplate files had to be fixed since they assumed the cursor was a global.
The hardcoded use of `^` was removed from there now that %ydboctocursor is a local variable.

In src/config.h and src/octo_init.c, raw_session and raw_schema fields were removed as they were
otherwise unused. raw_cursor was set to point to `cursor[0]` (instead of `cursor[1]`) now that there
is no `^` prefix.

Because ^%ydboctosession is no longer a global, a few tests that checked on this global were no longer
relevant so had to be removed.

In addition, `octo -k` and `rocto -k` as well as the `auto_clean_tables` option in the configuration file
no longer made sense since the cleanup happens always now that it is a local variable that goes away
when the process goes away. This meant changes to some documentation (doc/config.rst and doc/intro.rst)
as well as code (src/aux/octo.conf.default).

In src/rocto/handle_bind.c, a check for YDB_ERR_GVUNDEF was changed to YDB_ERR_LVUNDEF since the session
variable is now a local (not a global).