[#274] Octo now removes a cached table when it is recreated using CREATE TABLE
Octo will remove a table from its cache when it is recreated. This
allows the new definition to be active immediately. Previously it would
the cache would need to be cleared by relaunching octo or by using DROP
TABLE. Also includes a simplification of the DROP TABLE query to just
attempt to delete a cached table regardless of if it is in the cache or
not, and added test cases for this.