[#156] Added check for disabled null subscripts during Octo/Rocto initialization
Octo and Rocto require null subscripts to be enabled on their internal database and cannot function correctly otherwise. However, neither previously checked for this crucial setting and attempted to run even with it disabled.

Now, both Octo and Rocto check for this setting during initialization, notify the user, and exit gracefully. This is done with call-in to a simple M routine from the octo_init function.

Integration tests were added to verify behavior.