Reduce shift/reduce conflicts in parser (from 52 down to 28)
set_quantifier could take on an EMPTY value which meant that rules which were identical
except for the presence or absence of set_quantifier were ambiguous. Therefore the
rule which did not have the set_quantifier was removed. This made the new grammar equivalent
to the old one but avoided the shift/reduce conflicts.