Add configuration to auto start PostgreSQL and tests to prove it
Update dockerfiles to set the $USER variable to root as it is
undefined by default

Add fixtures for PostgreSQL to create tables and insert data

Add tests to verify that databases, tables, and data can be added
to PostgreSQL

Add special configuration files for CentOS as after the install of
PostgreSQL it needs more setup and is based on systemd. Use files
(see NOTICE) that change the systemd based setup to work with
docker containers

Update build.sh to start PostgreSQL and for CentOS configure it.