Rocto logs full SQL query when started with -vv (previously it used to strip out the last byte)
The issue was in parse_line where `cur_input_index` was decremented always. While it is okay if
we encounter a newline as the last character in the parse, it is not okay to do that otherwise.

Various existing test reference files have been fixed to account for the new code behavior.

A new testcase has been added to test_psql_connection.bats.in. Even though the issue existed only
in rocto, octo is also tested here just to be safe.