[#298] Added message length checks to read_flush, read_sync, and read_terminate
Previously, these functions didn't check messages read from the wire for length. Instead, they set the length to what it should be, without reporting the malformed message.

This behavior is now changed, so that messages are checked for the correct length per the PostgreSQL wire protocol specification. If an incorrect length is reaad, an ErrorResponse is generated, logged, and reported to client.