[#297] Do not compute cross-reference on a computed column (cannot establish a database trigger on data that does not exist)
The fix is in src/optimization_transforms/lp_generate_xref_keys.c where we check if EXTRACT was
specified for this column when the table was created. If so it is a computed column and so
do not try compute a cross reference.

A new test_computed_columns bats test has been added. The pre-existing namesWithAges table has been
enhanced to have a computed column.