[#237] Disable unconditional boolean expression expansion to Disjunctive Normal Form as it causes huge query plans
A prior commit (SHA 43a0b42e6a9bb32c5fd177ef891bb05f177a043f) had expanded/exploded boolean
statements unconditionally to allow for query optimization.

But this ended up being one of the major causes of performance slowdown when simple queries that
fetch the list of tables (or list of columns in a table) were run because they had a 3-way left join
coupled with a boolean expression that expanded to 22 terms when expanded to a disjunctive normal
form (see https://gitlab.com/YottaDB/DBMS/YDBOcto/issues/237 for query) caused a total of 27 *
22 = 594 plans to be generated (almost every row of the 3-way join gets accessed from the database
594 times). Not doing the boolean expansion would only cause 27 plans to be generated for the same
query. That means only 27 traversals of every row of the 3-way join from the database instead of
594 traversals).

Therefore the unconditional boolean expression expansion is disabled as part of this commit. This is
expected to speed up lots of queries in general.

The generated query plan (_ydboctoP*.m) file size for the #237 query used to be 58Mb before this
commit but is 1.6Mb after this change. This size can still be improved if we can optimize LEFT JOINs
to not create 3x plans but that will come as a separate change.

Note that this commit can slow down queries with simple boolean expressions involving OR. For
example a query like `select * from names where (id = 0 OR id = 1)` would perform faster when
the boolean expression is expanded as opposed to not since each expanded plan can be individually
optimized (e.g. id = 0) which results in just one row being accessed in each plan for a total of
two rows whereas the unexpanded query would go through all rows in the `names` table and evaluate
the boolean expression (id = 0 OR id = 1) for each of them. A separate issue
https://gitlab.com/YottaDB/DBMS/YDBOcto/issues/301 has been created for this.

A few existing test reference files have been fixed to reflect new output. In all of them, the actual
output is the same but the order is slightly different (because of a slightly different plan flow).
In some of them, we check that optimizations do kick in. Those have to be skipped/disabled currently
until the newly created issue #301 is fixed.

A new tests/test_boolean_expression_expansion.bats.in automated test case tests that the boolean
expression expansion does not happen like before.