[#237] Re-enable boolean expression expansion in case a max of ONE OUTER JOIN (LEFT/RIGHT/FULL) is used in the query
A recent commit (SHA 43a0b42e6a9bb32c5fd177ef891bb05f177a043f) disabled boolean expression expansion
unconditionally. It was noted in that commit that there are queries which might no longer run optimally
due to this change.

Turns out though that there are real-world queries with deep (e.g. 5-way) INNER JOIN usage which
will benefit from the boolean expression expansion particularly if the WHERE clause has expressions
which will help fix the value of a key.

In the below example query on a VistA database, there is a 5-way inner join whose performance
can be significantly improved by expanding the WHERE clause since it fixes A.status to be 5 or 6.
This causes us to only examine rows in the A table (ORDER1 table) which have a `status` value of 5
or 6.  This is expected to be a few hundreds or so compared to a million rows of the ORDER1 table
that we need to otherwise look at (if the boolean expression involving OR had not been expanded to
DNF/disjunctive-normal-form) before examining the WHERE clause for A.status value of 5 or 6.

```sql
SELECT  A.ORDER1_ID
FROM ORDER1 A
INNER JOIN ORDER_STATUS C ON (A.STATUS=C.ORDER_STATUS_ID)
INNER JOIN ORDER_ORDER_ACTIONS B ON (A.ORDER1_ID = B.ORDER1_ID AND A.CURRENT_ACTION=B.ORDER_ORDER_ACTIONS_ID)
INNER JOIN NEW_PERSON D ON (B.SIGNED_BY = D.NEW_PERSON_ID)
INNER JOIN PATIENT E ON (A.PATIENT_ID=E.PATIENT_ID)
INNER JOIN DISPLAY_GROUP F ON (A.TO=F.DISPLAY_GROUP_ID)
WHERE A.STATUS in (5,6)
ORDER by ORDER1_ID::integer
```

To make this work and at the same time not bloat queries involving outer joins and boolean
expressions, the boolean expression expansion to DNF is done only if there is at most one OUTER JOIN
(LEFT/RIGHT/FULL JOIN) in the query. This is an area of optimization that needs a lot more fine tuning
(will be done later as part of #301). For now, the immediate known performance issues are addressed.

A lot of pre-existing tests whose reference files were adjusted in 43a0b42e6a9bb32c5fd177ef891bb05f177a043f
to account for the changed behavior have now been reverted as they perform as well as before that commit
with this re-enabling of boolean expression expansion for the most commonly used cases.

A new test case has been added (TCKO2 subtest in test_composite_key bats test) to test the above
5-way INNER JOIN VistA query. The vista-mini.sql schema has been enhanced to add a few more tables that
are needed by the above query.