[#302] LVUNDEF error when SELECT column list involves a function invocation with a parameter from a LEFT JOIN table
The LVUNDEF error is because an occurrence of the table name was not replaced with NULL in case the table name
occurred as part of the SELECT column list which involved a function invocation where the tablename is used as
a parameter to the function. The function `replace_table_references()` takes care of this replacement but it
did not handle the case where a `value_STATEMENT` could in turn be a `CALCULATED_VALUE` which could in turn refer to
a table name that needs to be replaced with NULL. This is now fixed by addressing the `case value_STATEMENT:` block
in `replace_table_references.c` (and the similarly structured file `update_table_references.c`) like how it is handled
in `hash_canonical_query.c`.

In addition, an assert(FALSE) that was commented in hash_canonical_query() is now copied over to replace_table_references.c
and update_table_references.c with a comment indicating this needs to be removed once the open issue
https://gitlab.com/YottaDB/DBMS/YDBOcto/issues/171 is fixed.

While at this, the `set_operation_STATEMENT` case has also been handled in these two files like it is handled in
hash_canonical_query.c. This is necessary to fix https://gitlab.com/YottaDB/DBMS/YDBOcto/issues/305 but not
sufficient. More fixes are needed for that which will be done in a separate commit.

The pre-existing test_select_columns bats test has been enhanced to test this fix.