[#191] Added call to shutdown socket when Rocto encounters FATAL error
Previously, multiple SQL clients (PSQL and SquirrelSQL) would hang if Rocto encountered a FATAL error during query execution. This was due to Rocto server processes not closing their assigned sockets before exiting.

To resolve this, a call to shutdown was added to errors.c, allowing Rocto to close the socket before exiting.

A test was added to verify this behavior using the PSQL client. Though the original issue notes the erroneous behavior with SquirrelSQL, it was found to be reproducible with PSQL as well. Since PSQL can be easily scripted, it was chosen for testing purposes.