Optimize generated M code by avoiding redundant IF checks of keys that have already been fixed
If one runs the query `select * from names where id = 2`, the following M code is currently
generated in the physical plan (_ydboctoP*.m file).

```
  1 octoPlan0(cursorId)
  2     DO octoPlan1(cursorId)
  3     QUIT
  4
  5 octoPlan1(cursorId)
  6     NEW %ydboctoz KILL %ydboctocursor(cursorId,"keys",2,"","")
  7     SET %ydboctocursor(cursorId,"keys",0,"NAMES","ID")=+"2" IF $DATA(^names(...))'=0 DO
  8     . IF 1&(((%ydboctocursor(cursorId,"keys",0,"NAMES","ID")\1)=+"2")) DO
```

Notice that in line 7 we set %ydboctocursor(cursorId,"keys",0,"NAMES","ID") to the value 2 (due to
an optimization where we found an = check in the WHERE clause) but immediately afterwards in line
8, we have an IF check of %ydboctocursor(cursorId,"keys",0,"NAMES","ID")=2 (corresponding to the
same WHERE clause) which is redundant given what we did in line 7. This is because in the code, we
find out keys that can be fixed (lp_fix_key_to_const() function) based on the WHERE clause but do
not eliminate the fixed keys from the WHERE clause and that causes the generated M code to contain
these redundant checks in the IF condition.

This is now fixed by checking if lp_fix_key_to_const() returns TRUE and if so, removing this
LP_BOOLEAN_EQUALS part of the logical plan from the LP_WHERE boolean expression tree. This way
only the remaining relevant portions of the WHERE clause get generated in the M code. The fixed
keys are generated separately anyways and so it is okay to do this elimination of fixed keys.

This required changing the interface of lp_optimize_where_multi_equal_ands_helper() so we return
the input node in the expression tree if there is no elimination and return NULL if the input node
has been eliminated due to the fix-key-to-const optimization.

There is a subtlety with nested queries using column references from parent queries in the WHERE
clause which is solved by using where->v.operand[1] (currently unused) to maintain an alternate list
of optimized parent query references so the deferred plan determination (which currently relies
on the WHERE expression tree being untouched) works correctly even with the WHERE expression tree
modified due to elimination of various parts of the expression tree.

The existing test_where_optimizations bats test has been enhanced to do better verification of
the generated M code (it is now included as part of the reference file). And a few subtests that
were skipped as part of a recent commit (SHA fc8aa11bc5d5671ffd2e3a58f58072de70a56d63) but were
incorrectly not re-enabled as part of a later commit (that re-enabled boolan expression expansion)
are now re-enabled.

Similarly, the existing test_where_in bats test has also been enhanced. While verifying the
generated M code for TWI03 and TWI04 subtests, it was noticed that an equality check between a
column of the current query and a column from a parent query was not being fixed because of
an incorrect check in lp_optimize_where_multi_equal_ands_helper(). If key_unique_id_array[right_id]
is 0, it means the right column is coming in from a parent query which means left (current query column)
needs to be fixed to the value of right (parent query column) and so we should not swap them. We were
previously swapping them which meant we would later call lp_get_key() with the parent query column
which would return NULL and would later result in no call to lp_opt_fix_key_to_const() resulting in
no LP_FIX_KEY optimization. Verified that the newly generated M code correctly fixes the key.