[#270] LVUNDEF error when running a query involving a table with a composite key and a non-key-column fixed to a value
The main fix is to src/optimization_transforms/lp_optimize_where_multi_equals_ands.c where the case
a table can have multiple columns as the primary key (i.e. composite key) was not handled. The
previous code assumed there can be only one primary key column and this caused only the key for
the first of multiple columns to be added which meant that if the query accesses other columns in
the table which were defined to be based off other columns of the primary key then the generated M
code would reference the column/key but no FOR loop for that column/key would have been generated
before ending up in a LVUNDEF.

While at this, made some enhancements to src/optimization_transforms/generate_logical_plan.c and
src/optimization_transforms/lp_get_key.c which are aimed at improving the performance of queries that
involve joins with an ON clause and WHERE clause each of which help fix a particular column/key. In
that case, we would like to give preference to the ON clause as fixing that will help us better
speed up the query than fixing the WHERE clause.

Towards this, in generate_logical_plan.c, we add the ON clause (i.e. join_condition) in the join
order by invoking `lp_join_where()` so the current ON clause is appended to the ON clauses already
seen till now. And we add the WHERE clause after all ON clauses are added by invoking the later
`lp_join_where()` invocation with `where` as the second parameter instead of the first.

An existing test (TCD07) in the test_customers_database bats test failed due to an updated reference
file and is now fixed to reflect the updated output (new output computes a cross-reference for one
more column than before and looks at least not incorrect to me).

A new test_composite_key bats test has been introduced and a subtest in it checks that the LVUNDEF
does not happen in case a table has multiple primary key columns and a non-primary-key column is
fixed in the query. The query that resulted in the creation of #270 came from VistA SQL schema so
a mini version of that schema has been created in tests/fixtures/vista-mini.sql (no data has been
filled in so there is no tests/fixtures/vista-mini.zwr) and the new test as well a pre-existing
subtest (in the test_inner_join bats test) has been reworked to use this vista-mini.sql.